requirements.txt
python-telegram-bot==20.3
python-dotenv==1.0.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
aiofiles==23.2.1
requests==2.31.0
gspread
oauth2client

